Add in --dry-run (-n) flag

+++ src/bin/xm_decryptor.rs
@@ -1,9 +1,25 @@
 use std::path::PathBuf;
 
+use clap::Parser;
+
 use xm_decryptor::{xm, Result};
 
+/// Ximalaya xm file decryptor
+#[derive(Parser, Debug)]
+#[command(version, about, long_about = None)]
+struct Args {
+    /// Dry run?
+    #[arg(short = 'n', long, default_value_t = false)]
+    dry_run: bool,
+
+    #[clap(index = 1)]
+    pub xm_file: PathBuf,
+}
+
 fn main() -> Result<()> {
-    let path = PathBuf::from(std::env::args().nth(1).expect("no input path"));
+    let args = Args::parse();
+    // let path = PathBuf::from(std::env::args().nth(1).expect("no input path"));
+    let path = args.xm_file;
     let mut files = Vec::<PathBuf>::new();
     if path.is_file() {
         files.push(path);
@@ -21,14 +37,14 @@ fn main() -> Result<()> {
         .filter(|f| f.extension().unwrap_or_default() == "xm")
         .collect();
     for file in files {
-        if let Err(e) = decrypt_file(&file) {
+        if let Err(e) = decrypt_file(&file, args.dry_run) {
             eprintln!("error: {:?} {:?}", file, e);
         }
     }
     Ok(())
 }
 
-fn decrypt_file(file: &PathBuf) -> Result<()> {
+fn decrypt_file(file: &PathBuf, dry_run: bool) -> Result<()> {
     let content = std::fs::read(file)?;
 
     let xm_info = xm::extract_xm_info(&content[..])?;
@@ -39,6 +55,8 @@ fn decrypt_file(file: &PathBuf) -> Result<()> {
 
     let target_path = file.parent().expect("no parent dir").join(file_name);
     println!("target_path: {:?}", target_path);
-    std::fs::write(target_path, audio)?;
+    if !dry_run {
+        std::fs::write(target_path, audio)?;
+    }
     Ok(())
 }
